+++ app/src/app/about/page.tsx
@@ -1,13 +1,10 @@
-'use client'
 import AboutPage from "@/components/pages/about"
-import Introduce from "@/components/pages/about/introduce"
 
 export default function About() {
   
   return (
     <div>
       <AboutPage/>
-      <Introduce/>
     </div>
   )
 }
