+++ app/src/app/smartphonesite.tsx
@@ -1,0 +1,9 @@
+import Work from "@/components/pages/works";
+
+export default function WorkPage() {
+  return(
+    <div>
+      <Work/>
+    </div>
+  )
+}
