+++ app/src/components/pages/contact/hooks/index.ts
@@ -1,0 +1,11 @@
+type useContactProps = () => {
+  title: string
+}
+
+export const useContact: useContactProps = () => {
+  const title = "Contact"
+
+  return {
+    title
+  }
+}
