+++ app/src/components/pages/contact/presentations/contact-form.tsx
@@ -1,0 +1,10 @@
+export default function ContactForm() {
+  
+  return (
+    <div>
+      <div className="flex items-center justify-center">
+        <iframe src="https://docs.google.com/forms/d/e/1FAIpQLSf8W9oNnVZnIUaT38bXXp08QAsQsnBxY9z2EEUNB2TlUWRx3A/viewform?embedded=true" width="640" height="450">読み込んでいます…</iframe>
+      </div>
+    </div>
+  )
+}
