+++ app/src/components/pages/skill/index.tsx
@@ -1,12 +1,9 @@
-import Title from "@/components/title"
-export default function Contact() {
-  const title = "Contact"
-  return (
-    <div>
-      <Title name={title}/>
-      <div className="flex items-center justify-center">
-        <iframe src="https://docs.google.com/forms/d/e/1FAIpQLSf8W9oNnVZnIUaT38bXXp08QAsQsnBxY9z2EEUNB2TlUWRx3A/viewform?embedded=true" width="640" height="450">読み込んでいます…</iframe>
-      </div>
-    </div>
-  )
+import SkillList from '@/components/pages/skill/presentations/skill-list'
+
+export default function Skill() {
+	return (
+		<>
+			<SkillList />
+		</>
+	)
 }
