+++ app/src/libs/skill/skill-props.ts
@@ -1,0 +1,4 @@
+export interface SkillProps {
+  name: string
+  description: string
+}
