+++ app/src/libs/works/work-props.ts
@@ -1,0 +1,5 @@
+export interface WorkProps {
+  name: string
+  description: string
+  url: string | undefined
+}
